Add suites
A bit of boilerplate is required here. The openfx_sys suite
definitions define all fields as Option<fn(...)>. This is technically
correct since we can't guarantee that the function pointers are
non-null, but the ergonomics leave a lot to be desired.

Additionally, some arguments need conversion -- for example, we want
to use OfxStr where the low-level function takes *const c_char.

So we use a macro to generate wrappers for each suite function. The
wrappers check that arguments are valid, perform necessary
conversions, unwrap the function pointer, and call it.

In the future we could check that suite functions are non-null when
creating the suite wrappers, but that's not implemented yet.